Refactor logs to dedicated tab
